In preparation for 2.7.3.final
the gentoo subdir has been temporarily deleted, as it is
impractical to keep the ebuilds up-to-date, at the moment.

ebuild Manifests contain a checksum of the main
distribution tgz's, which are created at the same time
as the scala-tool-support sbaz package; the ebuild
Manifest regeneration should take place as part of the
release process.

For the moment the gentoo subdir is removed, pending a
better arrangement.



git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala-tool-support/trunk@16849 5e8d7ff9-d8ef-0310-90f0-a4852d11357a