Fix issue #46.
Rewrite scala-font-lock-limit function to narrow down the
"font-lock-multiline" marked text to actual multiple line construct. This
change largely decrease the size of text which need to be refontify when any
modification to text happen.

A minor change is made to function scala-match-and-skip-type-param to fix the
highlight problem for the code like below:

class A[T1,
        T2[T]
] (a: Int) { // "(a: Int)" will not be fontify as expected
}