Rewrite \` found in listing to a real \`
A \` that occurs in a listing is shown as two signs whereas one
wants to display it as a single `. This change is required because
a single ` that is written to a listing is already rewritten to and
therefore displayed as a beginning apostrophe.

This style bug can be seen in the syntax summary for `id`.