Wix.settings: mappings for creating the MSI.
Takes the result of staging universal and universal-docs,
and creates a WIX xml description for `packageBin in Windows`
to be able to create an MSI.

We no longer distributing examples, and link to GitHub for sources.

As .bat files are published to maven using the windows line endings,
no need to convert.

PS: Would've called in Windows.settings, but that clashes with the
packager's Windows config object.