Generic.settings: mappings for bin/, lib/, doc/ and man/
A Scala distribution is fully determined by the contents and dependencies of
the maven artifact "org.scala-lang" % "scala-dist" % version.value.

The contents of scala-dist itself is extracted to populate bin/, doc/ and man/.
The (transitive) dependencies populate the lib/ folder.

This configures sbt native packager's Universal config.

The next commit will take the jars under the javadoc classifier and extract/move
them to api/ to provide settings for the packager's UniversalDocs config.